redo fetch script to fetch single directory

+++ scripts/fetch.sh
@@ -1,9 +1,5 @@
 #!/bin/bash
-# Fetch your "labs" and "metacentrum" directory to your metacentrum storage.
-# When run fetches all the files from the remote
-# Then sends new files to the remote
-# Then tracks local files and updates them on the remote
-# Does not track file changes on the remote, restart is needed
+# Fetch directory from metacentrum
 
 # REQUIREMENTS (otherwise rsync will fail upon asking for password):
 #
@@ -22,13 +18,13 @@
 #
 # ssh-copy-id -i ~/.ssh/KEYNAME.pub USERNAME@tarkil.grid.cesnet.cz
 
-DIRS="./labs"
+if [[ ! -d $1 ]]; then
+   echo "Usage: $0 DIR"
+   exit 1
+fi
+DIR=$1
 
-# Start jobs
-for DIR in $DIRS; do
-    echo Fetching $DIR
-    rsync -azuptPE --delete metacentrum:"~/${DIR}/" "${DIR}/" 
-done
-
+echo Fetching $DIR
+rsync -azuptPE --delete metacentrum:"~/${DIR}/" "${DIR}/" 
 echo Done
